docker: chown installed files to root
This is to workaround the image not loading on DNAnexus. It fails with
the following message:

    ApplyLayer exit status 1 stdout:  stderr: lchown <path>: invalid argument

`<path>` is typically something that was installed, i.e.,
`/root/.cpanm` and `/opt/variants2matrix`.

There's something funny in the Ubuntu 16.04 (version 1) image on
DNAnexus, as I can't reproduce the error on a clean install of Ubuntu
16.04 and Docker 19.03.8.